requirements.txt
# langchain==0.1.14
# python-dotenv==1.0.0
# streamlit==1.22.0
# unstructured==0.9.2
# tiktoken==0.4.0
# faiss-cpu==1.7.4
# libmagic==1.0
# python-magic==0.4.27
# python-magic-bin==0.4.14
# google-generativeai==0.4.1  # Required for Gemini API
langchain==0.1.14
langchain-google-genai==0.0.9
streamlit==1.22.0
unstructured==0.9.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.3


